+++ CHANGELOG.md
@@ -5,9 +5,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [1.4.1]
+### Added
+- New hover visual for tile picker
+
 ### Fixed
 - You can now scroll all the way to the side when the width is maxed at 100% zoom
 - More robust tileset switcher
+- The tile picker shouldnt be outside the box anymore
 
 ## [1.4.0]
 ### Added